+++ tests/golden/ecom_basic/sqlmesh/duckdb/models/silver/order_item.sql
@@ -14,7 +14,7 @@ SELECT
     WHEN SUBSTRING(CAST(created_at AS TEXT), 11) LIKE '%+%'
     OR SUBSTRING(CAST(created_at AS TEXT), 11) LIKE '%-%'
     THEN NULL
-    ELSE created_at
+    ELSE REPLACE(REPLACE(CAST(created_at AS TEXT), 'T', ' '), 't', ' ')
   END AS TIMESTAMP) AT TIME ZONE 'Europe/Paris' AT TIME ZONE 'UTC' AS order_date,
   CAST(order_id AS TEXT) AS order_id,
   CAST(qty AS BIGINT) AS quantity,